Move gl stuff back into src/gl
