+++ src/gl/flamechart-color-pass-renderer.ts
