+++ src/gl/flamechart-renderer.ts
