+++ src/gl/overlay-rectangle-renderer.ts
