+++ src/gl/rectangle-batch-renderer.ts
