+++ src/views/flamechart-pan-zoom-view.tsx
@@ -1,8 +1,8 @@
 import {Rect, AffineTransform, Vec2, clamp} from '../lib/math'
 import {CallTreeNode} from '../lib/profile'
 import {Flamechart, FlamechartFrame} from '../lib/flamechart'
-import {CanvasContext} from '../gl2/canvas-context'
-import {FlamechartRenderer} from '../gl2/flamechart-renderer'
+import {CanvasContext} from '../gl/canvas-context'
+import {FlamechartRenderer} from '../gl/flamechart-renderer'
 import {Sizes, FontSize, Colors, FontFamily, commonStyle} from './style'
 import {cachedMeasureTextWidth, ELLIPSIS, trimTextMid} from '../lib/text-utils'
 import {style} from './flamechart-style'
